+++ src/IR/Domain/Process.hs
@@ -5,6 +5,9 @@ module IR.Domain.Process (
     ProcessName,
     mkProcessName,
     processNameText,
+    AppName,
+    mkAppName,
+    appNameText,
     IntervalMs,
     mkIntervalMs,
     intervalMilliseconds,
@@ -31,10 +34,22 @@ mkProcessName = mkName ProcessName
 processNameText :: ProcessName -> Text
 processNameText (ProcessName name) = name
 
+-- | FrogOS app/package identity as declared in package metadata.
+newtype AppName = AppName Text
+    deriving (Eq, Show)
+
+mkAppName :: Text -> Either DomainError AppName
+mkAppName = mkName AppName
+
+appNameText :: AppName -> Text
+appNameText (AppName name) = name
+
 -- | Process-domain conditions.
 data Condition
     = -- | True while the named process is present in the process table.
       ProcessRunning ProcessName (Maybe ObserveStrategy)
+    | -- | True while any process belonging to the named FrogOS app is running.
+      AppRunning AppName (Maybe ObserveStrategy)
     deriving (Eq, Show)
 
 attachStrategy :: ObserveStrategy -> Condition -> Either DomainError Condition
@@ -42,11 +57,19 @@ attachStrategy strategy (ProcessRunning name Nothing) =
     Right (ProcessRunning name (Just strategy))
 attachStrategy _strategy (ProcessRunning _name (Just _existingStrategy)) =
     Left StrategyConflict
+attachStrategy strategy (AppRunning name Nothing) =
+    Right (AppRunning name (Just strategy))
+attachStrategy _strategy (AppRunning _name (Just _existingStrategy)) =
+    Left StrategyConflict
 
 instance ToJSON Condition where
     toJSON (ProcessRunning (ProcessName n) obs) =
         object $
             ["type" .= ("process_running" :: Text), "name" .= n]
+                <> maybe [] (\s -> ["observe" .= s]) obs
+    toJSON (AppRunning (AppName n) obs) =
+        object $
+            ["type" .= ("app_running" :: Text), "name" .= n]
                 <> maybe [] (\s -> ["observe" .= s]) obs
 
 instance FromJSON Condition where
@@ -57,4 +80,8 @@ instance FromJSON Condition where
                 name <- o .: "name"
                 processName <- either (fail . T.unpack . renderDomainError) pure (mkProcessName name)
                 ProcessRunning processName <$> o .:? "observe"
+            "app_running" -> do
+                name <- o .: "name"
+                appName <- either (fail . T.unpack . renderDomainError) pure (mkAppName name)
+                AppRunning appName <$> o .:? "observe"
             _unknownType -> fail $ "unknown process condition: " <> T.unpack t